// src/types/config.ts
export type PresetColor =
  | "default"
  | "theme1"
  | "theme2"
  | "theme3"
  | "theme4"
  | "theme5"
  | "theme6";
export type I18n = "en" | "es" | "pt"; // 'en' - English, 'es' - Spanish, 'pt' - Portuguese

export enum MenuOrientation {
  VERTICAL = "vertical",
  HORIZONTAL = "horizontal",
}

export enum ThemeMode {
  LIGHT = "light",
  DARK = "dark",
}